+++ src/classes/job.ts
@@ -297,7 +297,7 @@ export class Job<
     }
   }
 
-  toJSON(): Omit<this, 'queue'> {
+  toJSON() {
     const { queue, ...withoutQueue } = this;
     return withoutQueue;
   }
@@ -855,7 +855,7 @@ export class Job<
   moveToWaitingChildren(
     token: string,
     opts: MoveToChildrenOpts = {},
-  ): Promise<boolean | Error> {
+  ): Promise<boolean> {
     return Scripts.moveToWaitingChildren(this.queue, this.id, token, opts);
   }